requirements.txt
blinker==1.8.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Flask==3.0.3
fonttools==4.54.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
tenacity==9.0.0
threadpoolctl==3.5.0
tzdata==2024.2
Werkzeug==3.0.4
